refactor: Split frontend.py into multiple files using APIRouters
fix: Fixed admin page routing
